`tools.microsoft_teams.get_user_id_by_email` -> `tools.microsoft_entra.get_user_id_by_email`
